bfitserver/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
click==8.1.7
Django==5.1.4
django-cors-headers==4.3.0
django-extensions==3.2.3
django-rq==3.0.0
djangorestframework==3.14.0
# psycopg2-binary==2.9.10
fpdf2==2.7.9
Jinja2==3.1.4
psycopg==3.2.4
psycopg-binary==3.2.4
pydicom==3.0.1
pytz==2024.2
redis==5.2.1
rq==2.0.0
requests==2.32.3
sqlparse==0.5.3
SimpleITK==2.4.0
typing_extensions==4.12.2
numpy==2.2.0
nibabel
connected-components-3d
pandas
scikit-image
watchdog
